+++ sockapi-ts/congestion/prologue.c
@@ -88,7 +88,7 @@ main(int argc, char *argv[])
     TEST_STEP("Increase @b RLIMIT_MEMLOCK up to @c SOCKTS_BPF_RLIMITS_MEMLOCK "
               "value on TST.");
     sockts_bpf_set_rlim_memlock(pco_tst,
-                                (unsigned int)SOCKTS_BPF_RLIMITS_MEMLOCK);
+                                (uint64_t)SOCKTS_BPF_RLIMITS_MEMLOCK);
 
     TEST_SUCCESS;
